# perf(http): make heap allocation for path conditional